AutoCleanAI/requirements.txt
pandas
scikit-learn
pandas-profiling
streamlit
apache-airflow
matplotlib
seaborn





aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.5.1
apache-airflow-providers-common-io==1.5.1
apache-airflow-providers-common-sql==1.24.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.1
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.1
apache-airflow-providers-sqlite==4.0.1
apispec==6.8.1
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncache==0.3.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.19.1
babel==2.17.0
bcrypt==4.1.2
blinker==1.9.0
boto3==1.34.56
botocore==1.34.162
bracex==2.6
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
curl_cffi==0.12.0
cycler==0.12.1
dacite==1.9.2
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dynaconf==3.2.6
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.7
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.56.0
forbiddenfruit==0.1.4
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-re2==1.1.20240702
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.26.0
ipywidgets==8.1.2
isodate==0.7.2
itables==1.7.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kr8s==0.13.5
lazy-object-proxy==1.10.0
limits==4.2
linkify-it-py==2.0.3
llvmlite==0.44.0
lockfile==0.12.2
loguru==0.7.3
Mako==1.3.9
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multimethod==1.12
mysql-connector-python==9.2.0
narwhals==1.30.0
networkx==3.4.2
nh3==0.2.17
numba==0.61.0
numpy==1.24.4
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-dbapi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlite3==0.50b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.50b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.2
pandas-profiling==3.6.6
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pendulum==3.0.0
phik==0.12.5
pid==3.0.4
pillow==11.1.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf==5.29.3
psutil==6.1.1
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
puremagic==1.30
py_fast_rsync==0.1.0
pyarrow==17.0.0
pycapnp==2.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-box==7.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jsonpath==1.3.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
RestrictedPython==7.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.23.1
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
setproctitle==1.3.5
shellingham==1.5.4
sherlock==0.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.5
streamlit==1.43.2
syft==0.9.5
syft-core==0.1.0
syft-event==0.1.0
syft-rpc==0.1.0
syftbox==0.3.5
sympy==1.14.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.5.0
text-unidecode==1.3
textual==3.7.1
threadpoolctl==3.6.0
time-machine==2.16.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typeguard==4.1.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
uvicorn==0.34.0
visions==0.8.1
watchdog==6.0.0
watchfiles==1.1.0
wcmatch==10.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wirerope==1.0.0
wordcloud==1.9.4
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
ydata-profiling==4.16.1
zipp==3.21.0
